refactor: extract RecordingRow and TagCountSlider to components/nav/
TagCloudModal imported both from RecordingsPanel, so drawing one row pulled in
the panel's whole 43-import graph. Moving them removes that back-edge; the
panel now exports nothing but its default.

RecordingRow takes MicIcon, CalendarIcon and RenameForm with it - one caller
each. Its useTranslation() stays deliberately un-namespaced: every key is
explicitly prefixed (workspace:/recordings:/common:) because the row draws on
all three and recordingMenu expects the same t. A "tidy" to
useTranslation("workspace") would break every label silently.

Pure move: RecordingsPanel.test.tsx is byte-identical and its 85 cases pass,
which is the no-regression gate. Panel 1380 -> 1068.
