+++ version.json
@@ -1,1 +1,1 @@
-{ "version": "0.186.5" }
+{ "version": "0.186.6" }
